[runtime] Use Int in ExternalPointer

+++ skipruntime-ts/addon/src/fromjs.cc
@@ -60,7 +60,7 @@ void* SkipRuntime_getContext() {
                                 "SkipRuntime_getContext", 0, nullptr);
 }
 
-CJArray SkipRuntime_Mapper__mapEntry(uint32_t mapperId, CJSON key,
+CJArray SkipRuntime_Mapper__mapEntry(int64_t mapperId, CJSON key,
                                      SKNonEmptyIterator values) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
@@ -73,7 +73,7 @@ CJArray SkipRuntime_Mapper__mapEntry(uint32_t mapperId, CJSON key,
                         "SkipRuntime_Mapper__mapEntry", 3, argv);
 }
 
-void SkipRuntime_deleteMapper(uint32_t mapperId) {
+void SkipRuntime_deleteMapper(int64_t mapperId) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
   Local<Value> argv[1] = {Number::New(isolate, mapperId)};
@@ -81,7 +81,7 @@ void SkipRuntime_deleteMapper(uint32_t mapperId) {
                      argv);
 }
 
-CJSON SkipRuntime_LazyCompute__compute(uint32_t lazyComputeId, char* self,
+CJSON SkipRuntime_LazyCompute__compute(int64_t lazyComputeId, char* self,
                                        CJSON key) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
@@ -94,7 +94,7 @@ CJSON SkipRuntime_LazyCompute__compute(uint32_t lazyComputeId, char* self,
                         "SkipRuntime_LazyCompute__compute", 3, argv);
 }
 
-void SkipRuntime_deleteLazyCompute(uint32_t lazyComputeId) {
+void SkipRuntime_deleteLazyCompute(int64_t lazyComputeId) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
   Local<Value> argv[1] = {Number::New(isolate, lazyComputeId)};
@@ -102,7 +102,7 @@ void SkipRuntime_deleteLazyCompute(uint32_t lazyComputeId) {
                      1, argv);
 }
 
-void SkipRuntime_ExternalService__subscribe(uint32_t externalSupplierId,
+void SkipRuntime_ExternalService__subscribe(int64_t externalSupplierId,
                                             char* collection, char* sessionId,
                                             char* resource, CJObject params) {
   Isolate* isolate = Isolate::GetCurrent();
@@ -118,7 +118,7 @@ void SkipRuntime_ExternalService__subscribe(uint32_t externalSupplierId,
                             "SkipRuntime_ExternalService__subscribe", 5, argv);
 }
 
-void SkipRuntime_ExternalService__unsubscribe(uint32_t externalSupplierId,
+void SkipRuntime_ExternalService__unsubscribe(int64_t externalSupplierId,
                                               char* sessionId) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
@@ -130,7 +130,7 @@ void SkipRuntime_ExternalService__unsubscribe(uint32_t externalSupplierId,
                      "SkipRuntime_ExternalService__unsubscribe", 2, argv);
 }
 
-double SkipRuntime_ExternalService__shutdown(uint32_t externalSupplierId) {
+double SkipRuntime_ExternalService__shutdown(int64_t externalSupplierId) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
   Local<Value> argv[1] = {Number::New(isolate, externalSupplierId)};
@@ -138,7 +138,7 @@ double SkipRuntime_ExternalService__shutdown(uint32_t externalSupplierId) {
                               "SkipRuntime_ExternalService__shutdown", 1, argv);
 }
 
-void SkipRuntime_deleteExternalService(uint32_t externalSupplierId) {
+void SkipRuntime_deleteExternalService(int64_t externalSupplierId) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
   Local<Value> argv[1] = {Number::New(isolate, externalSupplierId)};
@@ -146,7 +146,7 @@ void SkipRuntime_deleteExternalService(uint32_t externalSupplierId) {
                      "SkipRuntime_deleteExternalService", 1, argv);
 }
 
-char* SkipRuntime_Resource__instantiate(uint32_t resourceId,
+char* SkipRuntime_Resource__instantiate(int64_t resourceId,
                                         CJObject collections) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
@@ -158,7 +158,7 @@ char* SkipRuntime_Resource__instantiate(uint32_t resourceId,
                               "SkipRuntime_Resource__instantiate", 2, argv);
 }
 
-void SkipRuntime_deleteResource(uint32_t resourceId) {
+void SkipRuntime_deleteResource(int64_t resourceId) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
   Local<Value> argv[1] = {Number::New(isolate, resourceId)};
@@ -166,7 +166,7 @@ void SkipRuntime_deleteResource(uint32_t resourceId) {
                      argv);
 }
 
-SKResource SkipRuntime_ResourceBuilder__build(uint32_t builderId,
+SKResource SkipRuntime_ResourceBuilder__build(int64_t builderId,
                                               CJObject params) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
@@ -178,7 +178,7 @@ SKResource SkipRuntime_ResourceBuilder__build(uint32_t builderId,
                         "SkipRuntime_ResourceBuilder__build", 2, argv);
 }
 
-void SkipRuntime_deleteResourceBuilder(uint32_t resourceBuilderId) {
+void SkipRuntime_deleteResourceBuilder(int64_t resourceBuilderId) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
   Local<Value> argv[1] = {Number::New(isolate, resourceBuilderId)};
@@ -186,7 +186,7 @@ void SkipRuntime_deleteResourceBuilder(uint32_t resourceBuilderId) {
                      "SkipRuntime_deleteResourceBuilder", 1, argv);
 }
 
-void SkipRuntime_Checker__check(uint32_t executorId, char* request) {
+void SkipRuntime_Checker__check(int64_t executorId, char* request) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
   Local<Value> argv[2] = {
@@ -197,7 +197,7 @@ void SkipRuntime_Checker__check(uint32_t executorId, char* request) {
                      argv);
 }
 
-void SkipRuntime_deleteChecker(uint32_t checkerId) {
+void SkipRuntime_deleteChecker(int64_t checkerId) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
   Local<Value> argv[1] = {Number::New(isolate, checkerId)};
@@ -205,7 +205,7 @@ void SkipRuntime_deleteChecker(uint32_t checkerId) {
                      argv);
 }
 
-void SkipRuntime_Executor__resolve(uint32_t checkerId) {
+void SkipRuntime_Executor__resolve(int64_t checkerId) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
   Local<Value> argv[1] = {Number::New(isolate, checkerId)};
@@ -213,7 +213,7 @@ void SkipRuntime_Executor__resolve(uint32_t checkerId) {
                      1, argv);
 }
 
-void SkipRuntime_Executor__reject(uint32_t checkerId, double handle) {
+void SkipRuntime_Executor__reject(int64_t checkerId, double handle) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
   Local<Value> argv[2] = {Number::New(isolate, checkerId),
@@ -222,7 +222,7 @@ void SkipRuntime_Executor__reject(uint32_t checkerId, double handle) {
                      2, argv);
 }
 
-void SkipRuntime_deleteExecutor(uint32_t checkerId) {
+void SkipRuntime_deleteExecutor(int64_t checkerId) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
   Local<Value> argv[1] = {Number::New(isolate, checkerId)};
@@ -230,7 +230,7 @@ void SkipRuntime_deleteExecutor(uint32_t checkerId) {
                      argv);
 }
 
-void SkipRuntime_deleteService(uint32_t serviceId) {
+void SkipRuntime_deleteService(int64_t serviceId) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
   Local<Value> argv[1] = {Number::New(isolate, serviceId)};
@@ -238,7 +238,7 @@ void SkipRuntime_deleteService(uint32_t serviceId) {
                      argv);
 }
 
-CJObject SkipRuntime_Service__createGraph(uint32_t serviceId,
+CJObject SkipRuntime_Service__createGraph(int64_t serviceId,
                                           CJObject collections) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
@@ -250,7 +250,7 @@ CJObject SkipRuntime_Service__createGraph(uint32_t serviceId,
                         "SkipRuntime_Service__createGraph", 2, argv);
 }
 
-void SkipRuntime_Notifier__subscribed(uint32_t notifierId) {
+void SkipRuntime_Notifier__subscribed(int64_t notifierId) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
   Local<Value> argv[1] = {Number::New(isolate, notifierId)};
@@ -258,7 +258,7 @@ void SkipRuntime_Notifier__subscribed(uint32_t notifierId) {
                      "SkipRuntime_Notifier__subscribed", 1, argv);
 }
 
-void SkipRuntime_Notifier__notify(uint32_t notifierId, CJArray values,
+void SkipRuntime_Notifier__notify(int64_t notifierId, CJArray values,
                                   char* watermark, uint32_t updates) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
@@ -272,7 +272,7 @@ void SkipRuntime_Notifier__notify(uint32_t notifierId, CJArray values,
                      4, argv);
 }
 
-void SkipRuntime_Notifier__close(uint32_t notifierId) {
+void SkipRuntime_Notifier__close(int64_t notifierId) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
   Local<Value> argv[1] = {Number::New(isolate, notifierId)};
@@ -280,7 +280,7 @@ void SkipRuntime_Notifier__close(uint32_t notifierId) {
                      argv);
 }
 
-void SkipRuntime_deleteNotifier(uint32_t notifierId) {
+void SkipRuntime_deleteNotifier(int64_t notifierId) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
   Local<Value> argv[1] = {Number::New(isolate, notifierId)};
@@ -288,7 +288,7 @@ void SkipRuntime_deleteNotifier(uint32_t notifierId) {
                      argv);
 }
 
-CJSON SkipRuntime_Reducer__add(uint32_t reducerId, CJSON acc, CJSON value) {
+CJSON SkipRuntime_Reducer__add(int64_t reducerId, CJSON acc, CJSON value) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
   Local<Value> argv[3] = {
@@ -300,7 +300,7 @@ CJSON SkipRuntime_Reducer__add(uint32_t reducerId, CJSON acc, CJSON value) {
                         argv);
 }
 
-CJSON SkipRuntime_Reducer__remove(uint32_t reducerId, CJSON acc, CJSON value) {
+CJSON SkipRuntime_Reducer__remove(int64_t reducerId, CJSON acc, CJSON value) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
   Local<Value> argv[3] = {
@@ -312,7 +312,7 @@ CJSON SkipRuntime_Reducer__remove(uint32_t reducerId, CJSON acc, CJSON value) {
                                 "SkipRuntime_Reducer__remove", 3, argv);
 }
 
-void SkipRuntime_deleteReducer(uint32_t reducerId) {
+void SkipRuntime_deleteReducer(int64_t reducerId) {
   Isolate* isolate = Isolate::GetCurrent();
   Local<Object> externFunctions = kExternFunctions.Get(isolate);
   Local<Value> argv[1] = {Number::New(isolate, reducerId)};
